Restrict access to delete an article

+++ app/controllers/articles_controller.rb
@@ -48,9 +48,14 @@ before_action :set_article, only: [:show, :edit, :update, :destroy]
   end
 
   def destroy 
-    if @article.destroy
-      flash[:success] = "Article has been deleted"
-      redirect_to articles_path
+    unless @article.user == current_user 
+      flash[:danger] = "You cna only delete your own article."
+      redirect_to root_path
+    else
+      if @article.destroy
+        flash[:success] = "Article has been deleted"
+        redirect_to articles_path
+      end
     end
   end
